refactor: utils folder name 통일
