+++ packages/utils/src/index.ts
@@ -2,7 +2,7 @@ export * from './arrayUtils';
 export * from './assertion';
 export * from './colorUtils';
 export * from './functionUtils';
-export * from './localStorage';
+export * from './storageUtils';
 export * from './objectUtils';
 export * from './sharedUtils';
 export * from './strUtils';
